+++ src/platform.h
@@ -28,6 +28,7 @@ public:
     ~Mutex() { pthread_mutex_destroy(&mutex); }
     void lock() { pthread_mutex_lock(&mutex); }
     void unlock() { pthread_mutex_unlock(&mutex); }
+    bool trylock() { return pthread_mutex_trylock(&mutex) == 0; }
 private:
     friend class ConditionVariable;
     pthread_mutex_t mutex;